pyproject.toml
[tool.poetry]
name = "todo-api"
version = "0.1.0"
description = ""
authors = ["naveen-ravi2810 <rnaveen28102003@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.1"
python-decouple = "^3.8"
psycopg2 = "^2.9.9"
ruff = "^0.4.7"
bcrypt = "^4.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
